Update ArrayStatementFactory (#381)
